new_kingchess/net/requirements.txt
appdirs==1.4.4
Brotli==1.0.9
certifi==2024.8.30
charset-normalizer==2.0.4
cmake==3.30.2
coloredlogs==15.0.1
einops==0.8.0
filelock==3.13.1
flatbuffers==24.3.25
fsspec==2024.6.1
fvcore==0.1.5.post20221221
gmpy2==2.1.2
humanfriendly==10.0
idna==3.4
iopath==0.1.10
Jinja2==3.1.3
lit==18.1.8
Mako==1.3.2
MarkupSafe==2.1.3
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.23.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
onnx==1.17.0
packaging==24.1
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.0
portalocker==2.10.1
protobuf==5.26.0
pycuda==2024.1
PySocks==1.7.1
pytools==2024.1.1
PyYAML==6.0.2
requests==2.31.0
setuptools==68.2.2
sympy==1.12
tabulate==0.9.0
tensorrt==8.5.1.7
termcolor==2.4.0
torch==1.10.0
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.66.4
triton==2.0.0
typing_extensions==4.9.0
urllib3==2.1.0
wheel==0.41.2
yacs==0.1.8
